// frontend/src/api/model/DoctorPatientAppointmentDto.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * TeleHealth.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ConsultationSummaryDto } from "./ConsultationSummaryDto";
import type { LocalDate } from "./LocalDate";
import type { LocalTime } from "./LocalTime";
import type { Symptom } from "./Symptom";

export interface DoctorPatientAppointmentDto {
  publicId?: string;
  date?: LocalDate;
  startTime?: LocalTime;
  endTime?: LocalTime;
  visitReason?: string;
  status?: string;
  statusColorCode?: string;
  /** @nullable */
  symptoms?: Symptom[] | null;
  consultation?: null | ConsultationSummaryDto;
}