// frontend/src/api/model/PrescriptionSummaryDto.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * TeleHealth.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface PrescriptionSummaryDto {
  publicId?: string;
  medicationName?: string;
  dosage?: string;
  frequency?: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  durationDays?: number | string;
  takeWith?: string;
  warnings?: string[];
  storage?: string;
  missedDose?: string;
}